fold verifiers process helpers into existing modules
