+++ packages/prime/src/prime_lab_app/data.py
@@ -15,8 +15,8 @@ import toml
 from prime_cli.api.rl import RLClient
 from prime_cli.client import APIClient, APIError
 from prime_cli.core import Config
+from prime_cli.utils.eval_push import load_eval_config
 from prime_cli.utils.time_utils import format_time_ago
-from prime_cli.verifiers_process import load_eval_config
 from prime_evals import EvalsClient
 
 from .cache import (
